+++ examples/showoci/showoci_csv2adw.py
@@ -23,7 +23,12 @@
 # - OCI_SHOWOCI_BLOCK_VOLUMES
 # - OCI_SHOWOCI_DATABASE_SYSTEMS
 # - OCI_SHOWOCI_DATABASES_ADB
+# - OCI_SHOWOCI_FILE_STORAGE
+# - OCI_SHOWOCI_OBJECT_STORAGE
+# - OCI_SHOWOCI_LB_LISTENERS
+# - OCI_SHOWOCI_LB_BACKENDSET
 ##########################################################################
+
 import sys
 import argparse
 import datetime
@@ -32,7 +37,7 @@ import oracledb
 import time
 import os
 
-version = "23.02.14"
+version = "23.03.07"
 
 
 ##########################################################################
@@ -79,7 +84,7 @@ def get_time_elapsed(start_time):
 def set_parser_arguments():
 
     parser = argparse.ArgumentParser(formatter_class=lambda prog: argparse.HelpFormatter(prog, max_help_position=80, width=130))
-    parser.add_argument('-csv', default="", dest='csvlocation', help='CSV Location from showoci including header')
+    parser.add_argument('-csv', default="", dest='csv_location', help='CSV Location from showoci including header')
     parser.add_argument('-du', default="", dest='duser', help='ADB User')
     parser.add_argument('-dp', default="", dest='dpass', help='ADB Password')
     parser.add_argument('-dn', default="", dest='dname', help='ADB DSN')
@@ -88,11 +93,13 @@ def set_parser_arguments():
     parser.add_argument('-wl', default="", dest='wallet_location', help='Wallet Location')
     parser.add_argument('-wp', default="", dest='wallet_password', help='Wallet Password')
 
+    parser.add_argument('-drop', action='store_true', default=False, dest='drop', help='Drop Tables before Load')
+
     parser.add_argument('--version', action='version', version='%(prog)s ' + version)
 
     result = parser.parse_args()
 
-    if not (result.duser and result.dpass and result.dname and result.csvlocation):
+    if not (result.duser and result.dpass and result.dname and result.csv_location):
         parser.print_help()
         print_header("You must specify database credentials and csv location!!", 0)
         return None
@@ -103,7 +110,7 @@ def set_parser_arguments():
 ##########################################################################
 # Check Table Structure for Compute
 ##########################################################################
-def handle_compute(connection, csv_location):
+def handle_compute(connection, cmd):
     try:
 
         json = {
@@ -111,7 +118,7 @@ def handle_compute(connection, csv_location):
             'csv_file': "compute.csv",
             'items': [
                 {'col': 'tenant_name           ', 'csv': 'tenant_name           ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'tenant_id             ', 'csv': 'tenant_id             ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'tenant_id             ', 'csv': 'tenant_id             ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'instance_id           ', 'csv': 'instance_id           ', 'type': 'varchar2(1000)', 'pk': 'y'},
                 {'col': 'region_name           ', 'csv': 'region_name           ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'availability_domain   ', 'csv': 'availability_domain   ', 'type': 'varchar2(100) ', 'pk': 'n'},
@@ -136,7 +143,7 @@ def handle_compute(connection, csv_location):
                 {'col': 'boot_volume           ', 'csv': 'boot_volume           ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'boot_volume_size_gb   ', 'csv': 'boot_volume_size_gb   ', 'type': 'number        ', 'pk': 'n'},
                 {'col': 'boot_volume_b_policy  ', 'csv': 'boot_volume_b_policy  ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'boot_volume_encryption', 'csv': 'boot_volume_encryption', 'type': 'varchar2(20)  ', 'pk': 'n'},
+                {'col': 'boot_volume_encryption', 'csv': 'boot_volume_encryption', 'type': 'varchar2(200) ', 'pk': 'n'},
                 {'col': 'block_volumes         ', 'csv': 'block_volumes         ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'block_volumes_total_gb', 'csv': 'block_volumes_total_gb', 'type': 'number        ', 'pk': 'n'},
                 {'col': 'block_volumes_b_policy', 'csv': 'block_volumes_b_policy', 'type': 'varchar2(1000)', 'pk': 'n'},
@@ -145,7 +152,7 @@ def handle_compute(connection, csv_location):
                 {'col': 'extract_date          ', 'csv': 'extract_date          ', 'type': 'date          ', 'pk': 'n'}
             ]
         }
-        handle_table(connection, json, csv_location)
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
     except Exception as e:
         raise Exception("\nError at procedure: handle_compute - " + str(e))
 
@@ -153,7 +160,7 @@ def handle_compute(connection, csv_location):
 ##########################################################################
 # Check Table Structure for Block Storage
 ##########################################################################
-def handle_block_volume(connection, csv_location):
+def handle_block_volume(connection, cmd):
     try:
 
         json = {
@@ -161,7 +168,7 @@ def handle_block_volume(connection, csv_location):
             'csv_file': "block_volumes.csv",
             'items': [
                 {'col': 'tenant_name        ', 'csv': 'tenant_name        ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'tenant_id          ', 'csv': 'tenant_id          ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'tenant_id          ', 'csv': 'tenant_id          ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'id                 ', 'csv': 'id                 ', 'type': 'varchar2(1000)', 'pk': 'y'},
                 {'col': 'region_name        ', 'csv': 'region_name        ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'availability_domain', 'csv': 'availability_domain', 'type': 'varchar2(100) ', 'pk': 'n'},
@@ -172,14 +179,14 @@ def handle_block_volume(connection, csv_location):
                 {'col': 'backup_policy      ', 'csv': 'backup_policy      ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'vpus_per_gb        ', 'csv': 'vpus_per_gb        ', 'type': 'number        ', 'pk': 'n'},
                 {'col': 'volume_group_name  ', 'csv': 'volume_group_name  ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'instance_name      ', 'csv': 'instance_name      ', 'type': 'varchar2(100) ', 'pk': 'n'},
-                {'col': 'instance_id        ', 'csv': 'instance_id        ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'instance_name      ', 'csv': 'instance_name      ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'instance_id        ', 'csv': 'instance_id        ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'defined_tags       ', 'csv': 'defined_tags       ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'freeform_tags      ', 'csv': 'freeform_tags      ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'extract_date       ', 'csv': 'extract_date       ', 'type': 'date          ', 'pk': 'n'}
             ]
         }
-        handle_table(connection, json, csv_location)
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
     except Exception as e:
         raise Exception("\nError at procedure: handle_block_storage - " + str(e))
 
@@ -187,7 +194,7 @@ def handle_block_volume(connection, csv_location):
 ##########################################################################
 # Check Table Structure for Database All
 ##########################################################################
-def handle_database_systems(connection, csv_location):
+def handle_database_systems(connection, cmd):
     try:
 
         json = {
@@ -195,7 +202,7 @@ def handle_database_systems(connection, csv_location):
             'csv_file': "database_db_all.csv",
             'items': [
                 {'col': 'tenant_name         ', 'csv': 'tenant_name         ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'tenant_id           ', 'csv': 'tenant_id           ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'tenant_id           ', 'csv': 'tenant_id           ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'id                  ', 'csv': 'id                  ', 'type': 'varchar2(1000)', 'pk': 'y'},
                 {'col': 'region_name         ', 'csv': 'region_name         ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'availability_domain ', 'csv': 'availability_domain ', 'type': 'varchar2(100) ', 'pk': 'n'},
@@ -217,13 +224,13 @@ def handle_database_systems(connection, csv_location):
                 {'col': 'scan_ips            ', 'csv': 'scan_ips            ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'vip_ips             ', 'csv': 'vip_ips             ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'cluster_name        ', 'csv': 'cluster_name        ', 'type': 'varchar2(100) ', 'pk': 'n'},
-                {'col': 'domain              ', 'csv': 'domain              ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'domain              ', 'csv': 'domain              ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'db_nodes            ', 'csv': 'db_nodes            ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'db_homes            ', 'csv': 'db_homes            ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'maintenance_window  ', 'csv': 'maintenance_window  ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'last_maintenance_run', 'csv': 'last_maintenance_run', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'next_maintenance_run', 'csv': 'next_maintenance_run', 'type': 'varchar2(4000)', 'pk': 'n'},
-                {'col': 'infra_id            ', 'csv': 'infra_id            ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'infra_id            ', 'csv': 'infra_id            ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'cpu_core_count      ', 'csv': 'cpu_core_count      ', 'type': 'number        ', 'pk': 'n'},
                 {'col': 'shape_ocpus         ', 'csv': 'shape_ocpus         ', 'type': 'number        ', 'pk': 'n'},
                 {'col': 'db_storage_gb       ', 'csv': 'db_storage_gb       ', 'type': 'number        ', 'pk': 'n'},
@@ -236,7 +243,7 @@ def handle_database_systems(connection, csv_location):
                 {'col': 'extract_date        ', 'csv': 'extract_date        ', 'type': 'date          ', 'pk': 'n'}
             ]
         }
-        handle_table(connection, json, csv_location)
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
     except Exception as e:
         raise Exception("\nError at procedure: handle_database_systems - " + str(e))
 
@@ -244,7 +251,7 @@ def handle_database_systems(connection, csv_location):
 ##########################################################################
 # Check Table Structure for Databases
 ##########################################################################
-def handle_database_autonomous(connection, csv_location):
+def handle_database_autonomous(connection, cmd):
     try:
 
         json = {
@@ -252,7 +259,7 @@ def handle_database_autonomous(connection, csv_location):
             'csv_file': "database_autonomous.csv",
             'items': [
                 {'col': 'tenant_name            ', 'csv': 'tenant_name            ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'tenant_id              ', 'csv': 'tenant_id              ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'tenant_id              ', 'csv': 'tenant_id              ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'id                     ', 'csv': 'id                     ', 'type': 'varchar2(1000)', 'pk': 'y'},
                 {'col': 'region_name            ', 'csv': 'region_name            ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'compartment_path       ', 'csv': 'compartment_path       ', 'type': 'varchar2(2000)', 'pk': 'n'},
@@ -269,11 +276,11 @@ def handle_database_autonomous(connection, csv_location):
                 {'col': 'data_safe_status       ', 'csv': 'data_safe_status       ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'time_maintenance_begin ', 'csv': 'time_maintenance_begin ', 'type': 'varchar2(100) ', 'pk': 'n'},
                 {'col': 'time_maintenance_end   ', 'csv': 'time_maintenance_end   ', 'type': 'varchar2(100) ', 'pk': 'n'},
-                {'col': 'subnet_id              ', 'csv': 'subnet_id              ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'subnet_id              ', 'csv': 'subnet_id              ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'subnet_name            ', 'csv': 'subnet_name            ', 'type': 'varchar2(1000)', 'pk': 'n'},
                 {'col': 'private_endpoint       ', 'csv': 'private_endpoint       ', 'type': 'varchar2(1000)', 'pk': 'n'},
-                {'col': 'private_endpoint_label ', 'csv': 'private_endpoint_label ', 'type': 'varchar2(100) ', 'pk': 'n'},
-                {'col': 'nsg_ids                ', 'csv': 'nsg_ids                ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'private_endpoint_label ', 'csv': 'private_endpoint_label ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'nsg_ids                ', 'csv': 'nsg_ids                ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'nsg_names              ', 'csv': 'nsg_names              ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'whitelisted_ips        ', 'csv': 'whitelisted_ips        ', 'type': 'varchar2(4000)', 'pk': 'n'},
                 {'col': 'service_console_url    ', 'csv': 'service_console_url    ', 'type': 'varchar2(4000)', 'pk': 'n'},
@@ -288,9 +295,169 @@ def handle_database_autonomous(connection, csv_location):
                 {'col': 'extract_date           ', 'csv': 'extract_date           ', 'type': 'date          ', 'pk': 'n'}
             ]
         }
-        handle_table(connection, json, csv_location)
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
     except Exception as e:
         raise Exception("\nError at procedure: handle_database_autonomous - " + str(e))
+
+
+##########################################################################
+# Check Table Structure for File Storage
+##########################################################################
+def handle_file_storage(connection, cmd):
+    try:
+
+        json = {
+            'table_name': "OCI_SHOWOCI_FILE_STORAGE",
+            'csv_file': "file_storage.csv",
+            'items': [
+                {'col': 'tenant_name        ', 'csv': 'tenant_name        ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'tenant_id          ', 'csv': 'tenant_id          ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'id                 ', 'csv': 'id                 ', 'type': 'varchar2(1000)', 'pk': 'y'},
+                {'col': 'region_name        ', 'csv': 'region_name        ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'availability_domain', 'csv': 'availability_domain', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'compartment_path   ', 'csv': 'compartment_path   ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'compartment_name   ', 'csv': 'compartment_name   ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'display_name       ', 'csv': 'display_name       ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'size_gb            ', 'csv': 'size_gb            ', 'type': 'number        ', 'pk': 'n'},
+                {'col': 'exports            ', 'csv': 'exports            ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'snapshots          ', 'csv': 'snapshots          ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'mount_ips          ', 'csv': 'mount_ips          ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'defined_tags       ', 'csv': 'defined_tags       ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'freeform_tags      ', 'csv': 'freeform_tags      ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'time_created       ', 'csv': 'time_created       ', 'type': 'date          ', 'pk': 'n'},
+                {'col': 'extract_date       ', 'csv': 'extract_date       ', 'type': 'date          ', 'pk': 'n'}
+            ]
+        }
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
+    except Exception as e:
+        raise Exception("\nError at procedure: handle_file_storage - " + str(e))
+
+
+##########################################################################
+# Check Table Structure for File Storage
+##########################################################################
+def handle_object_storage(connection, cmd):
+    try:
+
+        json = {
+            'table_name': "OCI_SHOWOCI_OBJECT_STORAGE",
+            'csv_file': "object_storage_buckets.csv",
+            'items': [
+                {'col': 'tenant_name              ', 'csv': 'tenant_name              ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'tenant_id                ', 'csv': 'tenant_id                ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'id                       ', 'csv': 'bucket_id                ', 'type': 'varchar2(1000)', 'pk': 'y'},
+                {'col': 'namespace_name           ', 'csv': 'namespace_name           ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'region_name              ', 'csv': 'region_name              ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'compartment_path         ', 'csv': 'compartment_path         ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'compartment_name         ', 'csv': 'compartment_name         ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'bucket_name              ', 'csv': 'bucket_name              ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'size_gb                  ', 'csv': 'size_gb                  ', 'type': 'number        ', 'pk': 'n'},
+                {'col': 'objects                  ', 'csv': 'objects                  ', 'type': 'number        ', 'pk': 'n'},
+                {'col': 'object_lifecycle         ', 'csv': 'object_lifecycle         ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'public_access_type       ', 'csv': 'public_access_type       ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'storage_tier             ', 'csv': 'storage_tier             ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'is_read_only             ', 'csv': 'is_read_only             ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'versioning               ', 'csv': 'versioning               ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'auto_tiering             ', 'csv': 'auto_tiering             ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'kms_key_id               ', 'csv': 'kms_key_id               ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'logs                     ', 'csv': 'logs                     ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'defined_tags             ', 'csv': 'defined_tags             ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'freeform_tags            ', 'csv': 'freeform_tags            ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'preauthenticated_requests', 'csv': 'preauthenticated_requests', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'object_events_enabled    ', 'csv': 'object_events_enabled    ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'replication_enabled      ', 'csv': 'replication_enabled      ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'time_created             ', 'csv': 'time_created             ', 'type': 'date          ', 'pk': 'n'},
+                {'col': 'extract_date             ', 'csv': 'extract_date             ', 'type': 'date          ', 'pk': 'n'}
+            ]
+        }
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
+    except Exception as e:
+        raise Exception("\nError at procedure: handle_object_storage - " + str(e))
+
+
+##########################################################################
+# Check Table Structure for load_balancer_listeners
+##########################################################################
+def handle_load_balancer_listeners(connection, cmd):
+    try:
+
+        json = {
+            'table_name': "OCI_SHOWOCI_LB_LISTENERS",
+            'csv_file': "load_balancer_listeners.csv",
+            'items': [
+                {'col': 'tenant_name        ', 'csv': 'tenant_name        ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'tenant_id          ', 'csv': 'tenant_id          ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'id                 ', 'csv': 'id                 ', 'type': 'varchar2(1000)', 'pk': 'y'},
+                {'col': 'loadbalancer_id    ', 'csv': 'loadbalancer_id    ', 'type': 'varchar2(1000)', 'pk': 'y'},
+                {'col': 'region_name        ', 'csv': 'region_name        ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'compartment_path   ', 'csv': 'compartment_path   ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'compartment_name   ', 'csv': 'compartment_name   ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'name               ', 'csv': 'name               ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'status             ', 'csv': 'status             ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'shape              ', 'csv': 'shape              ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'type               ', 'csv': 'type               ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'ip_addresses       ', 'csv': 'ip_addresses       ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'log_errors         ', 'csv': 'log_errors         ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'log_access         ', 'csv': 'log_access         ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'logs               ', 'csv': 'logs               ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'subnets            ', 'csv': 'subnets            ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'listener_name      ', 'csv': 'listener_name      ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'listener_port      ', 'csv': 'listener_port      ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'listener_def_bs    ', 'csv': 'listener_def_bs    ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'listener_ssl       ', 'csv': 'listener_ssl       ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'listener_host      ', 'csv': 'listener_host      ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'listener_path      ', 'csv': 'listener_path      ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'listener_rule      ', 'csv': 'listener_rule      ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'lb_certificates    ', 'csv': 'lb_certificates    ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'defined_tags       ', 'csv': 'defined_tags       ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'freeform_tags      ', 'csv': 'freeform_tags      ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'time_created       ', 'csv': 'time_created       ', 'type': 'date          ', 'pk': 'n'},
+                {'col': 'extract_date       ', 'csv': 'extract_date       ', 'type': 'date          ', 'pk': 'n'}
+            ]
+        }
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
+    except Exception as e:
+        raise Exception("\nError at procedure: handle_load_balancer_listeners - " + str(e))
+
+
+##########################################################################
+# Check Table Structure for handle_load_balancer_backendset
+##########################################################################
+def handle_load_balancer_backendset(connection, cmd):
+    try:
+
+        json = {
+            'table_name': "OCI_SHOWOCI_LB_BACKENDSET",
+            'csv_file': "load_balancer_backendset.csv",
+            'items': [
+                {'col': 'tenant_name        ', 'csv': 'tenant_name        ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'tenant_id          ', 'csv': 'tenant_id          ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'id                 ', 'csv': 'id                 ', 'type': 'varchar2(1000)', 'pk': 'y'},
+                {'col': 'region_name        ', 'csv': 'region_name        ', 'type': 'varchar2(100) ', 'pk': 'n'},
+                {'col': 'compartment_path   ', 'csv': 'compartment_path   ', 'type': 'varchar2(2000)', 'pk': 'n'},
+                {'col': 'compartment_name   ', 'csv': 'compartment_name   ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'name               ', 'csv': 'name               ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'status             ', 'csv': 'status             ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'shape              ', 'csv': 'shape              ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'type               ', 'csv': 'type               ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'ip_addresses       ', 'csv': 'ip_addresses       ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'subnets            ', 'csv': 'subnets            ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'bs_name            ', 'csv': 'bs_name            ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'bs_desc            ', 'csv': 'bs_desc            ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'bs_status          ', 'csv': 'bs_status          ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'health_check       ', 'csv': 'health_check       ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'session_persistence', 'csv': 'session_persistence', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'ssl_cert           ', 'csv': 'ssl_cert           ', 'type': 'varchar2(1000)', 'pk': 'n'},
+                {'col': 'backend_name       ', 'csv': 'backend_name       ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'backend            ', 'csv': 'backend            ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'backend_ip         ', 'csv': 'backend_ip         ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'loadbalancer_id    ', 'csv': 'loadbalancer_id    ', 'type': 'varchar2(4000)', 'pk': 'n'},
+                {'col': 'extract_date       ', 'csv': 'extract_date       ', 'type': 'date          ', 'pk': 'n'}
+            ]
+        }
+        handle_table(connection, json, cmd.csv_location, cmd.drop)
+    except Exception as e:
+        raise Exception("\nError at procedure: handle_load_balancer_backendset - " + str(e))
 
 
 ##########################################################################
@@ -307,7 +474,7 @@ def variable_generation(item, index):
 ##########################################################################
 # Check Table Structure for Compute
 ##########################################################################
-def handle_table(connection, inputdata, csv_location):
+def handle_table(connection, inputdata, csv_location, drop_before_load):
     process_location = "Start"
     try:
 
@@ -343,6 +510,13 @@ def handle_table(connection, inputdata, csv_location):
             cursor.execute(sql, table_name=table_name)
             val, = cursor.fetchone()
 
+            # if main table exist and drop before load
+            if val > 0 and drop_before_load:
+                print("   Table " + table_name + " exist, but drop flag enabled, dropping..")
+                sql = "drop table " + table_name
+                cursor.execute(sql)
+                val = 0
+
             # if main table not exist, create it
             if val == 0:
                 print("   Table " + table_name + " was not exist, creating")
@@ -356,6 +530,13 @@ def handle_table(connection, inputdata, csv_location):
             sql = "select count(*) from user_tables where table_name = :table_name"
             cursor.execute(sql, table_name=tmp_table_name)
             val, = cursor.fetchone()
+
+            # if temp table exist and drop before load
+            if val > 0 and drop_before_load:
+                print("   Table " + tmp_table_name + " exist, but drop flag enabled, dropping..")
+                sql = "drop table " + tmp_table_name
+                cursor.execute(sql)
+                val = 0
 
             # if table not exist, create it
             if val == 0:
@@ -481,10 +662,14 @@ def main_process():
             print("   Connected")
 
             # Handling CSVs
-            handle_compute(connection, cmd.csvlocation)
-            handle_block_volume(connection, cmd.csvlocation)
-            handle_database_systems(connection, cmd.csvlocation)
-            handle_database_autonomous(connection, cmd.csvlocation)
+            handle_compute(connection, cmd)
+            handle_block_volume(connection, cmd)
+            handle_database_systems(connection, cmd)
+            handle_database_autonomous(connection, cmd)
+            handle_file_storage(connection, cmd)
+            handle_object_storage(connection, cmd)
+            handle_load_balancer_listeners(connection, cmd)
+            handle_load_balancer_backendset(connection, cmd)
 
     except oracledb.DatabaseError as e:
         print("\nError manipulating database - " + str(e) + "\n")